Introduce a new platform::get_info<>(info::platform) method
